# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur
    lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sistema Akuntansi API
  version: "1.0"
paths:
...
  /api/v1/monitoring/sessions/active:
    get:
      consumes:
      - application/json
      description: Get all active sessions for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Active sessions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active sessions
      tags:
      - Session Management
...
  /api/v1/sessions/cleanup:
    post:
      consumes:
      - application/json
      description: Manually trigger cleanup of expired sessions and tokens
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup completed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Force session cleanup
      tags:
      - Session Management
  /api/v1/sessions/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about active, expired, and total sessions
      produces:
      - application/json
      responses:
        "200":
          description: Session statistics
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get session statistics
      tags:
      - Session Management
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"